refactor: 모어체리 사이트 이미지 다운로드 기능 리팩토링
